completed2

+++ Binomial.java
@@ -19,8 +19,8 @@ return binomial1(n - 1, k) + binomial1(n - 1, k - 1);}
 	
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
-		int [][] memo = new int [n+1][k+1];
+	public static long binomial(int n, int k) {
+		long [][] memo = new long [n+1][k+1];
 		for (int i = 0 ; i<n+1 ; i++){
 			for ( int j = 0 ; j<k+1 ; j++){
 				memo[i][j] = -1;
@@ -30,7 +30,7 @@ return binomial1(n - 1, k) + binomial1(n - 1, k - 1);}
 		}
 	
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
